Add assembly level attribute
